# docs/swagger.yaml
definitions:
  socket_internal_dto.AuthBody:
    properties:
      password:
        minLength: 9
        type: string
      username:
        minLength: 4
        type: string
    required:
    - password
    - username
    type: object
  socket_internal_dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  socket_internal_dto.SuccessResponse-socket_internal_dto_UserResponse:
    properties:
      data:
        $ref: '#/definitions/socket_internal_dto.UserResponse'
    type: object
  socket_internal_dto.SuccessResponse-socket_internal_dto_UserWithTokenResponse:
    properties:
      data:
        $ref: '#/definitions/socket_internal_dto.UserWithTokenResponse'
    type: object
  socket_internal_dto.UserResponse:
    properties:
      username:
        type: string
    type: object
  socket_internal_dto.UserWithTokenResponse:
    properties:
      accessToken:
        type: string
      user:
        $ref: '#/definitions/socket_internal_dto.UserResponse'
    type: object
info:
  contact:
    name: API Support
  description: This is a swagger to show all RestAPI of chadChat project
  title: chadChat API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to the system
      parameters:
      - description: Description of the authentication body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/socket_internal_dto.AuthBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/socket_internal_dto.SuccessResponse-socket_internal_dto_UserWithTokenResponse'
        "400":
          description: your request is invalid or"your request body is incorrect or
            cannot save user
          schema:
            $ref: '#/definitions/socket_internal_dto.ErrorResponse'
        "500":
          description: cannot use this password
          schema:
            $ref: '#/definitions/socket_internal_dto.ErrorResponse'
      summary: Login to the system
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a user
      parameters:
      - description: Description of the authentication body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/socket_internal_dto.AuthBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/socket_internal_dto.SuccessResponse-socket_internal_dto_UserResponse'
        "400":
          description: your request is invalid or"your request body is incorrect or
            cannot save user
          schema:
            $ref: '#/definitions/socket_internal_dto.ErrorResponse'
        "409":
          description: username already exists
          schema:
            $ref: '#/definitions/socket_internal_dto.ErrorResponse'
        "500":
          description: cannot use this password
          schema:
            $ref: '#/definitions/socket_internal_dto.ErrorResponse'
      summary: Register a user
      tags:
      - Auth
swagger: "2.0"